Added NN lib, cleaned up and added examples
Added basic feedforward NN lib. Contains also few "tests" with
backpropagation and GA (GA not working though, but gives ideas on
how to use it. Removed  many things from the repo that *seemed*
unnecessary to me (i.e. they do not belong to the gir repo). If this
is not the case, then we have to see how to merge. I personally prefer
CMake as build system so I added that also.